REFACTOR: views, routes 분리 app.py 간소화
